+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
++ Support storing any type as a default context value [PR #136](https://github.com/shadaj/slinky/pull/136)
 
 ## v0.4.2
 + Fix bug with `shouldComponentUpdate` not being registered correctly on the component [PR #135](https://github.com/shadaj/slinky/pull/135)